chore: swicthing to monorepo formatter
