--- internal/node-patches/.gitignore
@@ -1,4 +1,0 @@
-coverage
-node_modules
-build
-.nyc_output
