--- internal/node-patches/.nycrc
@@ -1,1 +1,0 @@
-{"include":["build/src/**","src/**"]}
