+++ packages/node-patches/register.js
@@ -19,13 +19,12 @@
  */
 const patcher = require('./build/src/index.js');
 // todo auto detect bazel env vars instead of adding a new one.
-const { BAZEL_PATCH_ROOT, NP_SUBPROCESS_BIN_DIR, VERBOSE_LOGS } = process.env;
+const {BAZEL_PATCH_ROOT, NP_SUBPROCESS_BIN_DIR, VERBOSE_LOGS} = process.env;
 
 if (BAZEL_PATCH_ROOT) {
   if (VERBOSE_LOGS)
-    console.log(
-      `bazel node patches enabled. root: ${BAZEL_PATCH_ROOT} symlinks in this directory will not escape`
-    );
+    console.log(`bazel node patches enabled. root: ${
+        BAZEL_PATCH_ROOT} symlinks in this directory will not escape`);
   const fs = require('fs');
   patcher.fs(fs, BAZEL_PATCH_ROOT);
 } else if (VERBOSE_LOGS) {
