+++ packages/node-patches/test/fs/escape.ts
@@ -14,14 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { escapeFunction } from '../../src/fs';
 import * as assert from 'assert';
 import * as path from 'path';
+
+import {escapeFunction} from '../../src/fs';
 
 describe('escape function', () => {
   it('isOutPath is correct', () => {
     const root = '/a/b';
-    const { isOutPath } = escapeFunction(root);
+    const {isOutPath} = escapeFunction(root);
 
     assert.ok(isOutPath('/a'));
     assert.ok(isOutPath('/a/c/b'));
@@ -31,7 +32,7 @@ describe('escape function', () => {
 
   it('isEscape is correct', () => {
     const root = '/a/b';
-    const { isEscape } = escapeFunction(root);
+    const {isEscape} = escapeFunction(root);
 
     assert.ok(isEscape('/a/c/boop', '/a/b/l'));
     assert.ok(isEscape('/a/c/boop', '/a/b'));
@@ -42,7 +43,7 @@ describe('escape function', () => {
 
   it('isEscape handles relative paths', () => {
     const root = './a/b';
-    const { isEscape } = escapeFunction(root);
+    const {isEscape} = escapeFunction(root);
 
     assert.ok(isEscape('./a/c/boop', './a/b/l'));
     assert.ok(isEscape('./a/c/boop', './a/b'));
